Refatora skills de Guardião para Multiclasse
